+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @drivenets/eslint-plugin-design-system
 
+## 0.0.14
+
+### Patch Changes
+
+- 0901193: Support TypeScript 6.x
+
 ## 0.0.13
 
 ### Patch Changes